requirements.txt
Django>=4.1.4,<4.2
djangorestframework>=3.14,<3.15
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.25.1,<0.26

# For Authentication
djangorestframework-simplejwt>=5.2.2,<5.3
cryptography>=38.0.4,<38.1
django-cors-headers >= 3.10.0,<3.11

# For .env
python-dotenv>=0.19.0,<0.20